处理用户登录，以及用户新增，修改删除，编辑，获取等功能

+++ src/main/java/com/huhuto/demo/controller/UserController.java
@@ -10,11 +10,14 @@ import org.springframework.ui.ModelMap;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
 
+import java.text.SimpleDateFormat;
+import java.util.Date;
 import java.util.List;
 
 
 @CrossOrigin
 @Controller
+@RequestMapping("/users")
 public class UserController {
 
     private static final long EXPIRE_DATE=30*60*1000;
@@ -36,7 +39,7 @@ public class UserController {
             return "login";
         }
 
-//登录
+   //登录
     @RequestMapping(value = "/login",method = RequestMethod.POST)
     public String login(String username, String password,ModelMap modelMap){
         UserBean userBean = userService.user(username,password);
@@ -63,16 +66,16 @@ public class UserController {
     @ResponseBody
     public AjaxResult loginApi(String username, String password){
         UserBean userBean = userService.user(username,password);
+        userService.updateUserStatue(userBean.getId(),1);
         String token=tokenResult.token(username,password);
-        System.out.println("userBean"+userBean);
         ModelAndView md =new ModelAndView();
         if(userBean!=null){
             userService.updateToken(userBean.getId(),token);
-            md.addObject("username",userBean.getUsername());
-            md.addObject("password",userBean.getPassword());
-            md.addObject("id",userBean.getId());
-            md.addObject("userBean",userBean);
-            md.addObject("token",token);
+//            md.addObject("username",userBean.getUsername());
+//            md.addObject("password",userBean.getPassword());
+//            md.addObject("id",userBean.getId());
+//            md.addObject("userBean",userBean);
+//            md.addObject("token",token);
             return AjaxResult.successData(200,userBean);
         }else {
             return AjaxResult.error(400,"登录失败！！！");
@@ -80,16 +83,47 @@ public class UserController {
 
     }
 
+    //根据id查找用户信息
 
-//查询所有用户
-    @RequestMapping(value = "/allUser",method = RequestMethod.GET)
+    @RequestMapping(value = "/userById",method = RequestMethod.POST)
     @ResponseBody
-    public AjaxResult allUser(){
-        List<UserBean> userBean = userService.allUser();
+    public AjaxResult userById(@RequestBody UserBean userBean1){
+        UserBean userBean = userService.userById(userBean1.getId());
+
         System.out.println("userBean"+userBean);
+
         if(userBean!=null){
 
             return AjaxResult.successData(200,userBean);
+        }else {
+            return AjaxResult.error(400,"处理失败！！！");
+        }
+
+    }
+
+    //退出登录
+
+    @RequestMapping(value = "/logout",method = RequestMethod.POST)
+    @ResponseBody
+    public AjaxResult logout(@RequestBody UserBean userBean){
+
+        userService.updateUserStatue(userBean.getId(),0);
+       return AjaxResult.successData(200,"退出成功");
+
+    }
+
+
+
+//查询所有用户
+    @RequestMapping(value = "/allUser",method = RequestMethod.POST)
+    @ResponseBody
+    public AjaxResult allUser(int page, int limit,String sort,String title){
+        List<UserBean> userBean = userService.allUser(page,limit,sort,title);
+        System.out.println("userBean"+userBean);
+        if(userBean!=null){
+           int total=userBean.size();
+
+            return AjaxResult.successTableData(200,userBean,total);
         }else {
             return AjaxResult.error(400,"没有数据没有哦！！！");
         }
@@ -102,16 +136,30 @@ public class UserController {
     @RequestMapping(value = "/deleteUser",method = RequestMethod.POST)
     @ResponseBody
     public AjaxResult deleteUser(@RequestBody UserBean userBean){
+        boolean a = false;
+        UserBean userBean1 = userService.userById(userBean.getId());
 
-        System.out.println("userBean:"+userBean.getId());
-        boolean a = userService.deleteUser(userBean.getId());
+       if(!userBean1.getUsername().equals("admin")){
+           if(userBean1.getStatue()==0){
+               a = userService.deleteUser(userBean.getId());
+               if(a==true){
 
-        if(a==true){
+                   return AjaxResult.success("删除成功");
+               }else {
+                   return AjaxResult.error(400,"删除失败！！！");
+               }
+           }else{
 
-            return AjaxResult.success("删除成功");
-        }else {
-            return AjaxResult.error(400,"删除失败！！！");
-        }
+               return AjaxResult.error(400,"当前账户在线，删除失败！！！");
+
+           }
+       }else{
+
+
+           return AjaxResult.error(400,"管理员无法被删除！");
+       }
+
+
     }
 
 
@@ -140,6 +188,7 @@ public class UserController {
 
         System.out.println("userBean:"+userBean);
         boolean a = userService.updateUser(userBean);
+
 
         if(a==true){
